Version 1.13.2 Fixed various typos. Thanks to yayoKB for the help! Added check for MESSAGE_WRITE permission, because there are actually channels, that the bot can read, but not write in. :/ Added option to show a unique message in Random Facts on Purr's birthday (19th of march) Removed image from .server and added creation-date. Added CmdStats to show Stats (Discords, TextChannels, VoiceChannels, Members, ...) Added creation- and join-date to .user Replaced current image-thing for welcome-image with a own one from the ImageUtil.

+++ src/main/java/net/Andre601/commands/Info/CmdInfo.java
@@ -22,6 +22,9 @@ public class CmdInfo implements Command {
     @Override
     public void action(String[] args, MessageReceivedEvent e) {
 
+        if (!PermUtil.canWrite(e.getMessage()))
+            return;
+
         if(PermUtil.canDeleteMsg(e.getMessage()))
             e.getMessage().delete().queue();
 
@@ -36,7 +39,7 @@ public class CmdInfo implements Command {
                         "and a lot of free time. ;)\n" +
                         "\n" +
                         "**Commands**\n" +
-                        "You can use %shelp on the Discord, to see all my commands.",
+                        "You can use %shelp on your server to see all my commands.",
                         CmdPrefix.getPrefix(e.getGuild())
                 ))
                 .addField("Bot-Version:", String.format(
